@@ -1,5 +1,4 @@
 ﻿using AspNetCoreServerSide.Contracts;
-using AspNetCoreServerSide.Infrastructure;
 using AspNetCoreServerSide.Services;
 using AutoMapper;
 using JqueryDataTables.ServerSide.AspNetCoreWeb.DependencyInjection;
@@ -7,6 +6,8 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
+using Microsoft.Extensions.Hosting;
+using Newtonsoft.Json.Serialization;
 
 namespace AspNetCoreServerSide
 {
@@ -16,31 +17,37 @@ namespace AspNetCoreServerSide
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
         public void ConfigureServices(IServiceCollection services)
         {
-            services.AddScoped<IDemoService,DefaultDemoService>();
+            services.AddScoped<IDemoService, DefaultDemoService>();
 
             // Use in-memory database for quick dev and testing
             services.AddDbContext<Fingers10DbContext>(
-                options => {
+                options =>
+                {
                     options.UseInMemoryDatabase("fingers10db");
                 });
 
-            services.AddMvc();
+            services.AddControllersWithViews()
+                    .AddNewtonsoftJson(options => options.SerializerSettings.ContractResolver = new DefaultContractResolver())
+                    .AddJqueryDataTables();
             services.AddSession();
-            services.AddJqueryDataTables();
             services.AddAutoMapper(typeof(Startup));
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app,IHostingEnvironment env)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
-            if(env.IsDevelopment())
+            if (env.IsDevelopment())
             {
                 app.UseDeveloperExceptionPage();
             }
 
+            app.UseRouting();
             app.UseStaticFiles();
             app.UseSession();
-            app.UseMvcWithDefaultRoute();
+            app.UseEndpoints(endpoints =>
+            {
+                endpoints.MapControllerRoute("default", "{controller=Home}/{action=Index}/{id?}");
+            });
         }
     }
 }
