@@ -3,7 +3,7 @@ using Microsoft.EntityFrameworkCore;
 
 namespace AspNetCoreServerSide
 {
-    public class Fingers10DbContext:DbContext
+    public class Fingers10DbContext : DbContext
     {
         public Fingers10DbContext(DbContextOptions options)
             : base(options) { }
